requirements.txt
dash-core-components==1.12.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
gunicorn==20.1.0
matplotlib==3.1.1
numpy==1.17.2
paho-mqtt==1.5.0
pandas==0.25.1
psycopg2==2.8.4
scipy==1.3.1
